refactor(runtime): split builtins_atomic.rs into cohesive modules (§7-8)
`builtins_atomic.rs` was 1156 lines, over the 500-line module limit. Split it
into three files along its natural seams, each under 440 lines, with no behavior
change:

- `builtins_atomic.rs` (373) — atomic scalar-variable ops: native dispatch entry
  (`try_native_atomic_function`), name canonicalization, the shared-store
  value-key machinery, and the fetch/store/add/inc/dec primitives.
- `builtins_atomic_cas.rs` (373) — compare-and-swap on scalar variables
  (`cas($var, $expected, $new)` and the `cas($var, &code)` block form) and on
  single array elements.
- `builtins_atomic_shared.rs` (433) — thread-safe shared-container mutations
  (`shared_array_extend`/`shared_array_elem_set`/`shared_hash_elem_set`),
  multidim/hash-element CAS, the `multidim_get`/`multidim_set` navigators, and
  the `self_attr_cell_target` instance-attribute cell resolver.

`self_attr_cell_target` is widened from private to `pub(super)` since its callers
now span all three modules; `multidim_get`/`multidim_set` stay private with their
sole caller in `_shared`. The `ATOMIC_VAR_KEY_COUNTER` static stays with its only
user in `builtins_atomic`. Second slice of the §7-8 file-splitting sweep.
